requirements_inference.txt
torch==2.4.1
torchvision==0.19.1
accelerate==0.34.0
diffusers==0.24.0
einops==0.8.0
gradio==5.8.0
imageio==2.36.0
numpy==1.26.4
omegaconf==2.3.0
onnx==1.17.0
onnxruntime-gpu==1.20.1
onnx-graphsurgeon==0.5.2
opencv-python==4.10.0.84
pillow==11.0.0
pycuda==2024.1.2
tensorrt==10.1.0
transformers==4.30.2
matplotlib==3.9.3
huggingface-hub==0.25.1
polygraphy==0.49.9